feat(playpage): extract ReleaseLabel and simplify ReleaseSelector
Move rendering and label logic for each release into a new
ReleaseLabel component. Remove in-component installation status and
short-name logic from ReleaseSelector and import the shared hooks
and ReleaseLabel instead. This cleans up the ReleaseSelector by
reducing its responsibilities, improving readability, and making
label rendering reusable.

Also remove unused imports and selectors, update hook usage, and
adjust the useMemo dependencies accordingly.
